Replaced APIs + temporarily disabled code on  common and share extensions
