+++ Wire-iOS Share Extension/PreprocessorHelper.m
@@ -1,30 +1,40 @@
-// 
+//
 // Wire
 // Copyright (C) 2016 Wire Swiss GmbH
-// 
+//
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
 // the Free Software Foundation, either version 3 of the License, or
 // (at your option) any later version.
-// 
+//
 // This program is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 // GNU General Public License for more details.
-// 
+//
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see http://www.gnu.org/licenses/.
-// 
+//
 
 
-#import "AppDelegate.h"
+#import "PreprocessorHelper.h"
+@import Foundation;
 
-#import <HockeySDK/HockeySDK.h>
+#define STRINGIZE_INTERNAL(x) #x
+#define STRINGIZE(x) STRINGIZE_INTERNAL(x)
 
+NSString *wr_companyLoginURLScheme() {
+    return @STRINGIZE(WIRE_SSO_URL_SCHEME);
+}
 
+NSString * wr_appCenterAppId() {
+    return @STRINGIZE(APP_CENTER_APP_ID_KEY);
+}
 
-@interface AppDelegate (Hockey) <BITHockeyManagerDelegate, BITCrashManagerDelegate>
-
-- (void)setupHockeyWithCompletion:(dispatch_block_t)completed;
-
-@end
+BOOL wr_useAppCenter() {
+#if USE_APP_CENTER
+    return YES;
+#else
+    return NO;
+#endif
+}
