+++ Wire-iOS/Sources/Helpers/ConfigurationHelper.h
@@ -1,6 +1,6 @@
-////
+//
 // Wire
-// Copyright (C) 2019 Wire Swiss GmbH
+// Copyright (C) 2016 Wire Swiss GmbH
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -16,16 +16,8 @@
 // along with this program. If not, see http://www.gnu.org/licenses/.
 //
 
-#import <UIKit/UIKit.h>
+#import <Foundation/Foundation.h>
 
-//! Project version number for WireCommonComponents.
-FOUNDATION_EXPORT double WireCommonComponentsVersionNumber;
-
-//! Project version string for WireCommonComponents.
-FOUNDATION_EXPORT const unsigned char WireCommonComponentsVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <WireCommonComponents/PublicHeader.h>
-
-#import <WireCommonComponents/NetworkStatus.h>
-#import <WireCommonComponents/AVAsset+VideoConvert.h>
-#import <AppCenter/AppCenter.h>
+NSString *wr_companyLoginURLScheme(void);
+NSString * wr_appCenterAppId();
+BOOL wr_useAppCenter();
