+++ Wire-iOS/Sources/Helpers/ConfigurationHelper.m
